modified:   app/(tabs)/_layout.tsx 	modified:   app/(tabs)/friends/[userId].tsx 	modified:   app/(tabs)/friends/_layout.tsx 	modified:   app/(tabs)/profile/[userId].tsx 	modified:   app/(tabs)/profile/_layout.tsx 	modified:   app/(tabs)/profile/piggyBankModal.tsx 	modified:   app/(tabs)/profile/piggyBanks.tsx 	modified:   app/(tabs)/profile/wishListModal.tsx 	modified:   app/(tabs)/profile/wishModal.tsx 	modified:   app/(tabs)/profile/wishes.tsx 	modified:   app/(tabs)/settings/_layout.tsx 	modified:   app/(tabs)/settings/changeProfileBackground.tsx 	modified:   app/(tabs)/settings/index.tsx 	modified:   app/(tabs)/settings/pickProfileBackgroundColor.tsx 	modified:   components/ActionsButton.tsx 	modified:   components/BackButton.tsx 	modified:   components/FriendCard.tsx 	modified:   components/FullWishCard.tsx 	modified:   components/ProfileHeader.tsx 	modified:   components/TextInput.tsx 	modified:   components/WishCard.tsx 	renamed:    hooks/useProfile.tsx -> hooks/useStore.tsx 	modified:   hooks/useTheme.tsx 	modified:   models.ts 	modified:   utils/formatDate.ts 	new file:   utils/getDaysUntil.ts 	deleted:    utils/getDaysUntilBirthday.ts

+++ app/(tabs)/profile/[userId].tsx
@@ -1,5 +1,5 @@
-import { useCallback, useEffect, useRef, useState } from 'react';
-import { View, StyleSheet, Pressable, ScrollView, TouchableOpacity, RefreshControl } from 'react-native';
+import { useEffect, useRef, useState } from 'react';
+import { View, StyleSheet, Pressable, ScrollView, TouchableOpacity, RefreshControl, Alert } from 'react-native';
 import { ParallaxScrollView } from '@/components/ParallaxScrollView';
 import { ProfileHeader } from '@/components/ProfileHeader';
 import MasonryList from '@react-native-seoul/masonry-list';
@@ -11,15 +11,15 @@ import { Link, router, useLocalSearchParams } from 'expo-router';
 import { useTheme } from '@/hooks/useTheme';
 import Animated, { runOnJS, useAnimatedStyle, useSharedValue, withTiming } from 'react-native-reanimated';
 import { ThemedText } from '@/components/ThemedText';
-import { Booking, Friend, Profile, ProfileBackground, ServerProfileBackground, Wish, WishList } from '@/models';
+import { Booking, Friend, Profile, ProfileBackground, ApiProfileBackground, Wish, WishList } from '@/models';
 import { API } from '@/constants/api';
 import { Colors } from '@/constants/themes';
 import { ProgressBar } from '@/components/ProgressBar';
 import { apiFetchData, apiFetchImage } from '@/lib/api';
 import { useAuth } from '@/hooks/useAuth';
-import { useProfile } from '@/hooks/useProfile';
+import { useProfile } from '@/hooks/useStore';
 import { Action } from '@/components/ActionsButton';
-import { binaryArrayToBase64 } from '@/utils/convertImage';
+import { base64ToBinaryArray, binaryArrayToBase64 } from '@/utils/convertImage';
 import { getDefaultBackground } from '@/utils/profileBackground';
 
 type SearchParams = {
@@ -65,10 +65,10 @@ export default function ProfileScreen() {
 
   const [isRefreshing, setIsRefreshing] = useState(false);
 
-  const onRefresh = useCallback(() => {
+  const onRefresh = () => {
     setIsRefreshing(true);
     fetchData().then(() => setIsRefreshing(false));
-  }, []);
+  };
 
   const contentOpacity = useSharedValue(1);
   const contentAnimatedStyle = useAnimatedStyle(() => ({
@@ -112,7 +112,7 @@ export default function ProfileScreen() {
         token: authUser.token,
       }).then(setAvatar);
 
-      apiFetchData<ServerProfileBackground>({
+      apiFetchData<ApiProfileBackground>({
         endpoint: API.profile.getBackground(+userId),
         token: authUser.token,
       }).then((serverBackground) => {
@@ -130,37 +130,56 @@ export default function ProfileScreen() {
       const bookingsPromise = apiFetchData<Booking[]>({
         endpoint: API.profile.getBookings(+userId),
         token: authUser.token,
-      }).then((data) => {
+      }).then(async (data) => {
         setBookings(data);
-        data.forEach(async (booking) => {
-          const image = await apiFetchImage({
-            endpoint: API.wishes.getImage(booking.wish.wishId),
-            token: authUser.token,
-          });
-          setBookings((prev) =>
-            prev.map((prevBooking) =>
-              prevBooking.wish.wishId === booking.wish.wishId
-                ? { ...prevBooking, wish: { ...prevBooking.wish, image } }
-                : prevBooking
-            )
-          );
-        });
+
+        const imagesMap = new Map(
+          await Promise.all(
+            data.map(async (booking) => {
+              const image = await apiFetchImage({
+                endpoint: API.wishes.getImage(booking.wish.wishId),
+                token: authUser.token,
+              });
+              return [booking.wish.wishId, image] as const;
+            })
+          )
+        );
+
+        const updatedBookings = data.map((booking) => ({
+          ...booking,
+          wish: {
+            ...booking.wish,
+            image: imagesMap.get(booking.wish.wishId),
+          },
+        }));
+
+        setBookings(updatedBookings);
       });
 
       const wishesPromise = apiFetchData<Wish[]>({
         endpoint: API.profile.getWishes(+userId),
         token: authUser.token,
-      }).then((data) => {
+      }).then(async (data) => {
         setWishes(data);
-        data.forEach(async (wish) => {
-          const image = await apiFetchImage({
-            endpoint: API.wishes.getImage(wish.wishId),
-            token: authUser.token,
-          });
-          setWishes((prev) =>
-            prev.map((prevWish) => (prevWish.wishId === wish.wishId ? { ...prevWish, image } : prevWish))
-          );
-        });
+
+        const imagesMap = new Map(
+          await Promise.all(
+            data.map(async (wish) => {
+              const image = await apiFetchImage({
+                endpoint: API.wishes.getImage(wish.wishId),
+                token: authUser.token,
+              });
+              return [wish.wishId, image] as const;
+            })
+          )
+        );
+
+        const updatedWishes = data.map((wish) => ({
+          ...wish,
+          image: imagesMap.get(wish.wishId),
+        }));
+
+        setWishes(updatedWishes);
       });
 
       const wishListsPromise = apiFetchData<WishList[]>({
@@ -171,19 +190,27 @@ export default function ProfileScreen() {
       const piggyBanksPromise = apiFetchData<Wish[]>({
         endpoint: API.profile.getPiggyBanks(+userId),
         token: authUser.token,
-      }).then((data) => {
+      }).then(async (data) => {
         setPiggyBanks(data);
-        data.forEach(async (piggyBank) => {
-          const image = await apiFetchImage({
-            endpoint: API.wishes.getImage(piggyBank.wishId),
-            token: authUser.token,
-          });
-          setPiggyBanks((prev) =>
-            prev.map((prevPiggyBank) =>
-              prevPiggyBank.wishId === piggyBank.wishId ? { ...prevPiggyBank, image } : prevPiggyBank
-            )
-          );
-        });
+
+        const imagesMap = new Map(
+          await Promise.all(
+            data.map(async (piggyBank) => {
+              const image = await apiFetchImage({
+                endpoint: API.wishes.getImage(piggyBank.wishId),
+                token: authUser.token,
+              });
+              return [piggyBank.wishId, image] as const;
+            })
+          )
+        );
+
+        const updatedPiggyBanks = data.map((piggyBank) => ({
+          ...piggyBank,
+          image: imagesMap.get(piggyBank.wishId),
+        }));
+
+        setPiggyBanks(updatedPiggyBanks);
       });
 
       promises.push(bookingsPromise, wishesPromise, wishListsPromise, piggyBanksPromise);
@@ -199,17 +226,27 @@ export default function ProfileScreen() {
     const friendsPromise = apiFetchData<Friend[]>({
       endpoint: API.friends.getFriends(+userId),
       token: authUser.token,
-    }).then((data) => {
-      setFriends(data);
-      data.forEach(async (friend) => {
-        const avatar = await apiFetchImage({
-          endpoint: API.profile.getAvatar(friend.friendId),
-          token: authUser.token,
-        });
-        setFriends((prev) =>
-          prev.map((prevFriend) => (prevFriend.friendId === friend.friendId ? { ...prevFriend, avatar } : prevFriend))
-        );
-      });
+    }).then(async (friends) => {
+      setFriends(friends);
+
+      const avatarsMap = new Map(
+        await Promise.all(
+          friends.map(async (friend) => {
+            const avatar = await apiFetchImage({
+              endpoint: API.profile.getAvatar(friend.friendId),
+              token: authUser.token,
+            });
+            return [friend.friendId, avatar] as const;
+          })
+        )
+      );
+
+      const updatedFriends = friends.map((friend) => ({
+        ...friend,
+        avatar: avatarsMap.get(friend.friendId),
+      }));
+
+      setFriends(updatedFriends);
     });
 
     promises.push(profilePromise, friendsPromise);
@@ -236,9 +273,9 @@ export default function ProfileScreen() {
 
   const addItem = () => {
     if (currentTabIndex === 0) {
-      router.push('./wishModal');
+      router.push({ pathname: '/profile/wishModal' });
     } else if (currentTabIndex === 1) {
-      router.push('./piggyBankModal');
+      router.push({ pathname: '/profile/piggyBankModal' });
     }
   };
 
@@ -251,16 +288,25 @@ export default function ProfileScreen() {
   };
 
   const editWishList = (wishListId: number) => {
-    router.push({ pathname: './wishListModal', params: { wishListId } });
+    router.push({ pathname: '/profile/wishListModal', params: { wishListId } });
   };
 
   const deleteWishList = (wishListId: number) => {
-    apiFetchData({ endpoint: API.wishLists.delete(wishListId), method: 'DELETE', token: authUser.token })
-      .then(fetchMyWishLists)
-      .then(() => {
-        setCurrentWishListId(null);
-        scrollRef.current?.scrollTo({ x: 0, animated: true });
-      });
+    Alert.alert('Подтвердите', 'Вы уверены, что хотите удалить список?', [
+      { text: 'Отмена', style: 'cancel' },
+      {
+        text: 'Удалить список',
+        style: 'destructive',
+        onPress: () => {
+          apiFetchData({ endpoint: API.wishLists.delete(wishListId), method: 'DELETE', token: authUser.token })
+            .then(fetchMyWishLists)
+            .then(() => {
+              setCurrentWishListId(null);
+              scrollRef.current?.scrollTo({ x: 0, animated: true });
+            });
+        },
+      },
+    ]);
   };
 
   const shareWish = () => {
@@ -268,11 +314,23 @@ export default function ProfileScreen() {
   };
 
   const editWish = (wishId: number) => {
-    router.push({ pathname: './wishModal', params: { wishId } });
+    router.push({ pathname: '/profile/wishModal', params: { wishId } });
   };
 
   const deleteWish = (wishId: number) => {
-    apiFetchData({ endpoint: API.wishes.delete(wishId), method: 'DELETE', token: authUser.token }).then(fetchMyWishes);
+    Alert.alert('Подтвердите', 'Вы уверены, что хотите удалить желание?', [
+      { text: 'Отмена', style: 'cancel' },
+      {
+        text: 'Удалить желание',
+        style: 'destructive',
+        onPress: () => {
+          apiFetchData({ endpoint: API.wishes.delete(wishId), method: 'DELETE', token: authUser.token }).then(() => {
+            fetchMyWishes();
+            fetchMyWishLists();
+          });
+        },
+      },
+    ]);
   };
 
   const getBookingAction = (wishId: number): Action => {
@@ -295,13 +353,33 @@ export default function ProfileScreen() {
   };
 
   const editPiggyBank = (piggyBankId: number) => {
-    router.push({ pathname: './piggyBankModal', params: { piggyBankId } });
+    router.push({ pathname: '/profile/piggyBankModal', params: { piggyBankId } });
   };
 
   const deletePiggyBank = (piggyBankId: number) => {
     apiFetchData({ endpoint: API.wishes.delete(piggyBankId), method: 'DELETE', token: authUser.token }).then(
       fetchMyPiggyBanks
     );
+  };
+
+  const saveWish = (wish: Wish) => {
+    if (wish.image) {
+      apiFetchData({
+        endpoint: API.wishes.create,
+        method: 'POST',
+        body: {
+          wisherId: authUser.id,
+          wishType: 'TYPE_WISH',
+          name: wish.name,
+          description: wish.description,
+          price: wish.price,
+          currencyId: wish.currency?.currencyId,
+          link: wish.link,
+          image: base64ToBinaryArray(wish.image || ''),
+        },
+        token: authUser.token,
+      }).then(fetchMyWishes);
+    }
   };
 
   return (
@@ -382,7 +460,7 @@ export default function ProfileScreen() {
                     return (
                       <Link
                         asChild
-                        href={{ pathname: './wishes', params: { wishId: wish.wishId.toString() } }}
+                        href={{ pathname: '/profile/wishes', params: { wishId: wish.wishId, userId } }}
                         style={[
                           { marginTop: [0, 1].includes(i) ? 0 : 16 },
                           { [i % 2 === 0 ? 'marginRight' : 'marginLeft']: 8 },
@@ -402,7 +480,7 @@ export default function ProfileScreen() {
                                     { label: 'Удалить', onPress: () => deleteWish(wish.wishId) },
                                   ]
                                 : [
-                                    { label: 'Сохранить к себе', onPress: () => editWish(wish.wishId) },
+                                    { label: 'Сохранить к себе', onPress: () => saveWish(wish) },
                                     isFriend(+userId) ? getBookingAction(wish.wishId) : null,
                                     { label: 'Поделиться', onPress: shareWish },
                                   ].filter((action) => !!action)
